mv config files in config.py
