+++ deal/__init__.py
@@ -1,4 +1,5 @@
-from .core import DataConfig, DEALConfig, FlareConfig, DEAL
+from .config import DataConfig, DEALConfig, FlareConfig
+from .core import DEAL
 from .utils import *
 
 __all__ = ["DataConfig", "DEALConfig", "FlareConfig", "DEAL"]
